Added config system
Many bugfixes and codebase changes, some small features

+++ src/main/java/dev/thingymabobs/mixin/LevelTransformer.java
@@ -3,5 +3,5 @@ package dev.thingymabobs.mixin;
 import com.simibubi.create.content.redstone.link.IRedstoneLinkable;
 
 public interface LevelTransformer {
-	public int transform(int level, IRedstoneLinkable self, IRedstoneLinkable other);
+	public float transform(float level, IRedstoneLinkable self, IRedstoneLinkable other);
 }
